requirements.txt
aioping
dnspython